pyproject.toml
[project]
name = "memory_mcp"
version = "0.1.0"
description = ""
readme = "README.md"
authors = [
    { name = "Taylor Agarwal", email = "taagarwa@redhat.com" }
]
requires-python = ">=3.12"
dependencies = [
    "fastmcp>=3.1.0",
    "langchain>=1.2.10",
    "langgraph-checkpoint-postgres>=3.0.4",
    "psycopg[binary,pool]>=3.3.3",
]

[project.scripts]
main = "memory_mcp.main:main"

[dependency-groups]
dev = [
    "pytest~=8.1",
    "pytest-cov>=4.0.0,<5",
    "ruff>=0.4.1,<0.5",
    "bump2version>=1.0.1,<2",
    # "git-cliff>=2.8,<3",
]

[[tool.uv.index]]
name = "de-cop-nexus"
url = "https://nexus.corp.redhat.com/repository/de-cop-pypi-releases/simple"
explicit = true

[tool.ruff]
line-length = 140

[tool.ruff.lint]
select = ["E", "F", "D", "C", "N"]
ignore = [
    "E501", # line-too-long
    "E402", # module-import-not-at-top-file
    "D203", # one-blank-line-before-class
    "D212", # multi-line-summary-first-line
    "D100", # undocumented-public-module
]
per-file-ignores = { "tests/*" = ["D"] }

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
